Refactored and renamed 🏗️
