+++ app/[lang]/page.tsx
@@ -1,5 +1,5 @@
 import parse from 'html-react-parser'
-import Socials from '@/components/Socials'
+import Socials from '@/components/shared/social-icons'
 import Navbar from '@/components/navbar'
 import { getDictionary } from '@/get-dictionaries'
 import { Locale } from '@/i18n-config'
